// src/app/models/viewmodels/create-asset-stock-taking-detail.ts
export interface CreateAssetStockTakingDetail {
    assetId: string;
    assetStockTakingOrganizationId: string;
    responsibilityIdentity: string;
    responsibilityName: string;
    responsibilityOrg2: string;
    assetStockTakingLocation: string;
    stockTakingStatus: number;
    message: string;
}